Use QueryContext to hold DeleteManager
This holds a map of state that connectors can use to hold any state they
need for the duration of the query. Currently this is only used by the
ConnectorDynamicFilterProvider but it could be used by any other query.

In the context of IcebergDynamicFilterProvider it is used to store the
DeleteManager of the query to share delete files across different splits
of the query.