IcebergPlugin: Performance improvements for Equality Delete files
The goal of this commit is to improve the performance of queries that
contain Eqaulity Delete files.

Before this commit equality delete files were re-loaded for every split
that needed them. Now, each delete file is loaded once per execution
node.

In addition, the equality deletes are stored in a single map (per delete
schema) with the data sequence number in which the row was deleted. This
allows us to merge rows that were deleted multiple times (common in
upsert use cases) into a single entry instead of holding an entry per
delete file.

Changes in this commit:
* Create a DeleteManager class to manage the above logic
* Read every delete file only once
* Merge deletes into single map
* Opportunistic delete file load parallelization

State Management:
The `StatefulIcebergPageSourceProvider` is created per task and allows
for reusing delete file infomation within the task.
The `DeleteManager` is created only once and then loaded equality deletes for the
task are cached until the end of the task.

Tasks are per partition and typcially most equality delete files in a partition
apply to most data files so caching all the equality delete files is
efficient.